fix: createBooks_broken.js
In my earlier commit I mistakenly "fixed" a global variable `CSVFilePath`

The `<input id="CSVFilePath" …>` (in `createBooks.html`) makes its contents available (apparently) in a global variable `CSVFilePath`. I mistakenly thought it was a typo and changed it to `this.CSVFilePath` which is what broke it from functioning in the browser (it did pass the tests though).

Sorry about that. That was the only breaking piece of code. I've tested the functionality in browser and it correctly consumes the csv file and displays the resulting book structure.